-Added support for 64bits RHEL5 build machines. -On 64bits machine modules will be cross-built for 32bits machine too.
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/motor@286 ececa60f-9f63-0410-97c7-ea05888eb260

+++ motorApp/NewportSrc/devMM4000.cc
@@ -115,7 +115,7 @@ static struct board_stat **MM4000_cards;
 STATIC long MM4000_init(void *arg)
 {
     long rtnval;
-    int after = (int) arg;
+    long after = (long) arg;
 
     if (after == 0)
     {
